reorganize library folder
