+++ lib/src/features/library/presentation/library_view.dart
@@ -5,13 +5,13 @@ import 'package:book_adapter/src/features/in_app_update/util/toast_utils.dart';
 import 'package:book_adapter/src/features/library/data/book_collection.dart';
 import 'package:book_adapter/src/features/library/data/collection_section.dart';
 import 'package:book_adapter/src/features/library/data/item.dart';
-import 'package:book_adapter/src/features/library/library_view_controller.dart';
-import 'package:book_adapter/src/features/library/widgets/add_book_button.dart';
-import 'package:book_adapter/src/features/library/widgets/add_to_collection_button.dart';
-import 'package:book_adapter/src/features/library/widgets/merge_to_series.dart';
-import 'package:book_adapter/src/features/library/widgets/overflow_library_appbar_popup_menu_button.dart';
-import 'package:book_adapter/src/features/library/widgets/profile_button.dart';
-import 'package:book_adapter/src/features/library/widgets/section_widget.dart';
+import 'package:book_adapter/src/features/library/presentation/library_view_controller.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/add_book_button.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/add_to_collection_button.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/merge_to_series.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/overflow_library_appbar_popup_menu_button.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/profile_button.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/section_widget.dart';
 import 'package:book_adapter/src/localization/app.i18n.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
