+++ lib/src/features/library/presentation/series_view.dart
@@ -4,9 +4,9 @@ import 'package:book_adapter/src/constants/constants.dart';
 import 'package:book_adapter/src/features/in_app_update/util/toast_utils.dart';
 import 'package:book_adapter/src/features/library/data/book_item.dart';
 import 'package:book_adapter/src/features/library/data/series_item.dart';
-import 'package:book_adapter/src/features/library/library_view_controller.dart';
-import 'package:book_adapter/src/features/library/widgets/item_list_tile_widget.dart';
-import 'package:book_adapter/src/features/library/widgets/overflow_library_appbar_popup_menu_button.dart';
+import 'package:book_adapter/src/features/library/presentation/library_view_controller.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/item_list_tile_widget.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/overflow_library_appbar_popup_menu_button.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
