+++ lib/src/features/library/presentation/widgets/add_new_collection_button.dart
