+++ lib/src/features/library/presentation/widgets/default_item_image.dart
