+++ lib/src/features/library/presentation/widgets/delete_button.dart
