+++ lib/src/features/library/presentation/widgets/section_widget.dart
@@ -2,8 +2,8 @@ import 'package:book_adapter/src/constants/constants.dart';
 import 'package:book_adapter/src/features/in_app_update/util/toast_utils.dart';
 import 'package:book_adapter/src/features/library/data/collection_section.dart';
 import 'package:book_adapter/src/features/library/data/item.dart';
-import 'package:book_adapter/src/features/library/library_view_controller.dart';
-import 'package:book_adapter/src/features/library/widgets/item_list_tile_widget.dart';
+import 'package:book_adapter/src/features/library/presentation/library_view_controller.dart';
+import 'package:book_adapter/src/features/library/presentation/widgets/item_list_tile_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:implicitly_animated_reorderable_list/implicitly_animated_reorderable_list.dart';
